requirements.txt
# requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.29.0
pydantic==2.5.2
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.9
firebase-admin==6.2.0
python-dotenv==1.0.0
email-validator==2.1.0
pytz==2023.3
pytest==7.3.1
httpx==0.24.1
pydantic-settings==2.1.0
